refactor code to simplify
