+++ lib/glific/providers/gupshup/instrumentation.ex
@@ -4,13 +4,13 @@ defmodule Glific.Providers.Gupshup.Instrumentation do
 
   Inherits the standard provider counters (`track_send/2`, `track_receive/2`,
   `track_status/2`, `track_hsm_sync/2`) from
-  `Glific.Providers.Instrumentation.Adapter`, and adds Gupshup's one bit of
+  `Glific.Providers.Instrumentation`, and adds Gupshup's one bit of
   custom classification: a frequency-capped 4xx is recorded under a
   `frequency_capped` status rather than `error`, so throttled sends don't trip
   send-failure alerts.
   """
 
-  use Glific.Providers.Instrumentation.Adapter, provider: "gupshup"
+  use Glific.Providers.Instrumentation, provider: "gupshup"
 
   # NOTE: 472 is Gupshup's documented "Frequency Cap" code. The exact code(s) to
   # exclude are an open question on the monitoring ticket — confirm against live
@@ -18,12 +18,11 @@ defmodule Glific.Providers.Gupshup.Instrumentation do
   # untouched.
   @frequency_cap_error_codes [472]
 
-  @impl true
   def classify_send(:error, context) do
     if frequency_capped?(context[:error_code]), do: :frequency_capped, else: :error
   end
 
-  def classify_send(status, _context), do: status
+  def classify_send(status, context), do: super(status, context)
 
   @doc """
   Whether a Gupshup error code represents a frequency-capped send. Accepts the
